Fix date string length and experience height

+++ components/Experience/ExperienceOverview.tsx
@@ -10,7 +10,7 @@ export default function ExperienceOverview(experience: ExperienceType) {
       <h3 className="font-mohave text-6xl">{experience.attributes.company}</h3>
       <h4 className="text-3xl font-light">{experience.attributes.title}</h4>
       <h6 className="text-md font-light opacity-80">
-        {startDate.toLocaleDateString('default', { month: 'long', year: "numeric" })} - {endDate.toLocaleDateString('default', { month: 'short', year: "numeric" })}
+        {startDate.toLocaleDateString('default', { month: 'short', year: "numeric" })} - {endDate.toLocaleDateString('default', { month: 'short', year: "numeric" })}
       </h6>
       <h5 className="mt-4">{experience.attributes.summary}</h5>
